qdrant_implementation/requirements.txt
Flask
Flask-Limiter==3.9.2
gunicorn==23.0.0
numpy
openai==1.57.4
pandas
python-dotenv
requests
qdrant-client